Move code to api/file
